requirements.txt
Flask==2.3.2
twilio==8.0.0
gspread==5.11.0
oauth2client
python-dotenv==1.0.0